Adding separated out controller view of authorList and authorPage
