+++ src/components/authors/authorPage.js
@@ -2,6 +2,7 @@
 
 var React = require('react');
 var AuthorApi = require('../../api/authorApi');
+var AuthorList = require('./authorList');
 
 var Authors = React.createClass({
     getInitialState: function(){
@@ -10,33 +11,18 @@ var Authors = React.createClass({
         }
     },
 
-    componentWillMount: function(){
-        this.setState({ authors: AuthorApi.getAllAuthors() });
+    componentDidMount: function(){
+        if (this.isMounted()){
+            this.setState({ authors: AuthorApi.getAllAuthors() });
+        }
     },
 
     render: function(){
-        var createAuthorRow = function(author){
-            return (
-                <tr key={author.id}>
-                    <td><a href={"/#authors/" + author.id}>{author.id}</a></td>
-                    <td>{author.firstName} {author.lastName}</td>
-                </tr>
-            )
-        };
 
         return (
             <div className="container-fluid">
                 <h1>Authors</h1>
-                <table className="table">
-                    <thead>
-                        <th>ID</th>
-                        <th>Name</th>
-                    </thead>
-                    <tbody>
-                        {this.state.authors.map(createAuthorRow, this)}
-                    </tbody>
-
-                </table>
+                <AuthorList authors={this.state.authors} />
             </div>
         );
     }
